removing more static_cast and fixing consequences

+++ include/geos/noding/BasicSegmentString.h
@@ -92,7 +92,7 @@ public:
 	 *        Must not be the last index in the vertex list
 	 * @return the octant of the segment at the vertex
 	 */
-	int getSegmentOctant(unsigned int index) const;
+	int getSegmentOctant(size_t index) const;
 
 private:
